move hmmalign into its own function
